+++ www/Library/views/Focuser/Focuser.php
@@ -38,6 +38,7 @@ class Focuser extends Widget
         $object = null;
         // объект по умолчанию
         if (empty($uri) && ($object = Data::read(array(
+                'select' => 'children',
                 'from' => '/Contents',
                 'where' => array(
                     array('is', '/Library/content_samples/Page')
@@ -45,8 +46,7 @@ class Focuser extends Widget
                 'order' => array(
                     array('order', 'ASC')
                 ),
-                'limit' => array(0,1),
-                'depth' => 1
+                'limit' => array(0,1)
             )))){
             $object = reset($object);
         }