+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 2.0.alpha work in progress
 - new: Added support of conditions of string type (VladimirShestakov)
 - enh: Data::select() method  is combined with Data::read() (VladimirShestakov)
 - enh: Added style for output of function trace() (VladimirShestakov)
+- enh #123: Search of parents, prototypes and heirs is added. Syntax of the condition for search of children, counts of objects and check of existence of object is changed. Initial and final depth of search is added. (VladimirShestakov)
 
 Version 2.0.preview-2 April, 27 2013
 ----------------------------